pyproject.toml
[project]
name = "django-saintrelion-libs"
version = "3.0.5"
description = "Modular Django libraries"
dependencies = [
    "django",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-cors-headers",
    "webauthn",
    "django-otp",
    "celery"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"